Move util methods from BOOST_AUTO_TEST_SUITE
